requirements.txt
Django==1.10.3
Markdown==2.6.7
dj-database-url==0.4.1
django-bootstrap3==7.1.0
django-chartjs==1.1
django-filter==0.15.3
djangorestframework==3.5.3
gunicorn==19.6.0
psycopg2==2.6.2
pyotp==2.2.1
django-cors-headers==1.3.1
whitenoise==3.2.2
social-auth-app-django==0.0.1
pytz==2016.10